iio: jesd204: axi_adxcvr: Add TX|RX_PROGDIV support
This is useful for JESD204_ENCODER_64B66B modes.
Set TXOUTCLK LR/33 for GTH and LR/66 for GTH there will by another
DIV2 in the fabric which brings the LR/33 also to LR/66.
We always report this clock to be LR/66. This difference is handled
by setting the PROGDIV_RATE for GTY to 2.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>